edit to run on VM env

+++ stop.sh
@@ -5,6 +5,15 @@
 #    exit 1
 #fi
 
-jobid=`cat jobid`
-echo "running qdel $jobid"
-qdel $jobid
+
+if [ $ENV == "IUHPC" ]; then
+    jobid=`cat jobid`
+    echo "running qdel $jobid"
+    qdel $jobid
+fi
+
+if [ $ENV == "VM" ]; then
+    pid=`cat pid`
+    echo "running kill" 
+    kill $pid
+fi
